aider: Added function to retrieve a reviewer by their address.
# Aider chat conversation:

USER: create a function `getReviewer`

ASSISTANT: To create a new function `getReviewer` that retrieves a reviewer by their address, we need to modify the `contracts/PeerReview.sol` file. This function will search through the list of reviewers and return the details of the reviewer if the address matches.

Here's how we can implement this:

1. **File to modify**: `contracts/PeerReview.sol`
2. **Changes**:
   - Add a new function `getReviewerByAddress` that takes an address as input and returns the reviewer's details if found.

Here is the *SEARCH/REPLACE block* for adding the `getReviewerByAddress` function:

contracts/PeerReview.sol
```solidity
<<<<<<< SEARCH
    // Function to get a reviewer's information by index
    function getReviewer(uint256 index)
        public
        view
        returns (address, string[] memory)
    {
        Reviewer storage reviewer = reviewers[index];
        return (reviewer.addr, reviewer.keywords);
    }
=======
    // Function to get a reviewer's information by index
    function getReviewer(uint256 index)
        public
        view
        returns (address, string[] memory)
    {
        Reviewer storage reviewer = reviewers[index];
        return (reviewer.addr, reviewer.keywords);
    }

    // Function to get a reviewer's information by address
    function getReviewerByAddress(address reviewerAddress)
        public
        view
        returns (address, string[] memory)
    {
        for (uint256 i = 0; i < reviewers.length; i++) {
            if (reviewers[i].addr == reviewerAddress) {
                return (reviewers[i].addr, reviewers[i].keywords);
            }
        }
        revert("Reviewer not found.");
    }
>>>>>>> REPLACE
```

This function iterates through the `reviewers` array to find a reviewer by their address. If the reviewer is found, it returns the reviewer's address and keywords. If not found, it reverts with the message "Reviewer not found."